Make klp-convert locally compilable
klp-convert depends on macros which are defined within the Linux kernel source
code. To prevent this dependency, this patch makes it possible to define these
macros locally through a special make option (make local). This is useful for
testing purposes, but should not be used for building packages to be distributed
as, for functioning, it is required that these macros match.

Reviewed-by: Petr Mladek <pmladek@suse.com>
Signed-off-by: Joao Moreira <jmoreira@suse.de>